requirements.txt
pandas==2.1.0
PyYAML==6.0.1
numpy==1.25.2